requirements.txt
streamlit
graphviz
python-docx